+++ R/RPPA.R
@@ -50,6 +50,8 @@
 #' - 21Q1: no change, no further releases are scheduled at this time.
 #' 
 #' - 21Q2: no change, no further releases are scheduled at this time.
+#' 
+#' - 21Q3: no change, no further releases are scheduled at this time.
 #'
 #' @docType data
 #'